add  xALOAD / xASTORE instructions

+++ tests/P1001/Main.java
@@ -3,10 +3,10 @@ import java.util.*;
 public class Main {
     public static void main(String[] args) {
         Scanner cin = new Scanner(System.in);
-        Integer a[] = new Integer[3];
-        a[0] = cin.nextInt();
-        a[1] = cin.nextInt();
-        a[2] = a[0] + a[1];
+        short a[] = new short[3];
+        a[0] = cin.nextShort();
+        a[1] = cin.nextShort();
+        a[2] = (short)(a[0] + a[1]);
         System.out.println(a[2]);
     }
 }
